+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/src/app/pages/unknown-route/UnknownRoutePage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import useEffectOnce from '@navikt/sif-common-core-ds/lib/hooks/useEffectOnce';
+import { useEffectOnce } from '@navikt/sif-common-hooks';
 import { ErrorPage, SoknadErrorMessages } from '@navikt/sif-common-soknad-ds';
 import appSentryLogger from '../../utils/appSentryLogger';