+++ apps/sif-ettersending/src/app/types/Søknadstype.ts
@@ -1,11 +1,9 @@
-export enum ApplicationType {
-    'ukjent' = 'ukjent',
-    'pleiepengerBarn' = 'pleiepenger',
-    'pleiepengerLivetsSluttfase' = 'pleiepenger-livets-sluttfase',
+export enum Søknadstype {
+    'pleiepengerSyktBarn' = 'pleiepenger',
+    'pleiepengerLivetsSluttfase' = 'pleiepengerLivetsSluttfase',
     'omsorgspenger' = 'omsorgspenger',
     'ekstraomsorgsdager' = 'ekstraomsorgsdager',
     'utbetaling' = 'utbetaling',
     'utbetalingarbeidstaker' = 'utbetalingarbeidstaker',
     'regnetsomalene' = 'regnetsomalene',
-    'deling' = 'deling',
 }