+++ apps/sif-ettersending/src/app/utils/navigationUtils.ts
@@ -1,13 +1,13 @@
 import { NavigateFunction } from 'react-router-dom';
 import { getEnvironmentVariable } from '@navikt/sif-common-core-ds/lib/utils/envUtils';
-import { getRouteConfig, getAbsoluteUrlForRoute } from '../config/routeConfig';
-import { ApplicationType } from '../types/ApplicationType';
+import { getAbsoluteUrlForRoute, getRouteConfig } from '../config/routeConfig';
+import { Søknadstype } from '../types/Søknadstype';
 
-const getLoginUrl = (søknadstype: ApplicationType) => `${getEnvironmentVariable('LOGIN_URL')}/${søknadstype}`;
+const getLoginUrl = (søknadstype: Søknadstype) => `${getEnvironmentVariable('LOGIN_URL')}/${søknadstype}`;
 
 export const redirectTo = (route: string) => window.location.assign(route);
 export const navigateTo = (route: string, navigate: NavigateFunction) => navigate(route);
-export const navigateToErrorPage = (søknadstype: ApplicationType, navigate?: NavigateFunction) => {
+export const navigateToErrorPage = (søknadstype: Søknadstype, navigate?: NavigateFunction) => {
     const routeConfig = getRouteConfig(søknadstype);
     if (navigate) {
         navigate(routeConfig.ERROR_PAGE_ROUTE);
@@ -16,13 +16,13 @@ export const navigateToErrorPage = (søknadstype: ApplicationType, navigate?: Na
     }
 };
 
-export const navigateToKvitteringPage = (søknadstype: ApplicationType, navigate: NavigateFunction) =>
+export const navigateToKvitteringPage = (søknadstype: Søknadstype, navigate: NavigateFunction) =>
     navigate(getRouteConfig(søknadstype).APPLICATION_SENDT_ROUTE);
 
-export const navigateToLoginPage = (søknadstype: ApplicationType) => window.location.assign(getLoginUrl(søknadstype));
-export const navigateToWelcomePage = (søknadstype: ApplicationType) =>
+export const navigateToLoginPage = (søknadstype: Søknadstype) => window.location.assign(getLoginUrl(søknadstype));
+export const navigateToWelcomePage = (søknadstype: Søknadstype) =>
     window.location.assign(getAbsoluteUrlForRoute(getRouteConfig(søknadstype).WELCOMING_PAGE_ROUTE));
-export const userIsCurrentlyOnErrorPage = (søknadstype: ApplicationType) =>
+export const userIsCurrentlyOnErrorPage = (søknadstype: Søknadstype) =>
     window.location.pathname === getAbsoluteUrlForRoute(getRouteConfig(søknadstype).ERROR_PAGE_ROUTE);
 
 export const relocateToNavFrontpage = (): void => redirectTo('https://www.nav.no/');