+++ packages/sif-common-hooks/src/hooks/useHasChanged.ts
@@ -4,7 +4,6 @@ import { usePrevious } from './usePrevious';
 export const useHasChanged = (value: any, callback: (value: any) => void) => {
     const previous = usePrevious(value);
     const hasChanged = previous !== value;
-
     useEffect(() => {
         if (hasChanged) {
             callback && callback(previous);