+++ apps/pleiepenger-sykt-barn/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @navikt/pleiepenger-sykt-barn
 
+## 3.0.2
+
+### Patch Changes
+
+-   Pakkeoppdateringer
+-   Updated dependencies [55e1f323]
+-   Updated dependencies
+    -   @navikt/sif-common-core-ds@9.7.0
+    -   @navikt/sif-common-amplitude@2.21.1
+    -   @navikt/sif-common-formik-ds@0.6.1
+    -   @navikt/sif-common-soknad-ds@10.7.1
+    -   @navikt/sif-common-forms-ds@8.8.1
+    -   @navikt/appstatus-react-ds@2.14.1
+    -   @navikt/sif-common-sentry@0.22.1
+    -   @navikt/sif-common-utils@3.40.1
+
 ## 3.0.1
 
 ### Patch Changes